Split updater lifecycle types by ownership
