+++ Packages/macOS/CmuxUpdater/Sources/CmuxUpdater/UpdatePromptReply.swift
@@ -1,18 +1,6 @@
 import Foundation
 @preconcurrency public import Sparkle
 
-/// Why cmux consumed a Sparkle update prompt.
-///
-/// Sparkle's later `dismissUpdateInstallation` callback does not identify the prompt or the
-/// action that caused it. Recording the cause at the reply boundary lets the lifecycle owner
-/// distinguish an explicit user choice from a controller transition or accepted install.
-@MainActor
-enum UpdatePromptReplySource: String {
-    case user
-    case superseded
-    case installAttempt
-}
-
 /// A Sparkle prompt reply that can be sent at most once, with its consumption observable.
 ///
 /// Sparkle's update-choice callback must be invoked exactly once per prompt; double-replying is
